Use SQLite for the backing store
Cache::FastMmap was convenient, but I forgot one of its most important
characteristics: it's a _cache_, and so will evict entries when it runs
out of page space. This results in lost metrics if you have enough of
them.

So, use SQLite instead. It's a bit less fun to work with, but it will at
least store everything. Also it seems to be a bit faster, which is
perhaps not surprising but was not quite what I expected either.