Reconnect to database after fork
SQLite handles can't be reused after fork. This makes some attempt to
detect and reconnect to the database when this happens.

The included test doesn't really exercise this very well (it passed
before). The SQLite docs mostly say that the two sources of pain are
locks and cleanup, neither of which are easy to exercise. So hopefully
this is good enough.